+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fixed `.meta.json` file being picked as a script's file path instead of the actual Luau file
 - Fixed diagnostics not clearing for files when workspace diagnostics is not enabled
+- Fixed metatable name not being used when hovering over the function of a metatable
 
 ## [1.10.1] - 2022-09-24