+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added support for Semantic Tokens
 
+### Changed
+
+- Improved autocomplete items ordering by applying heuristics to sort items
+
 ### Fixed
 
 - Fixed `.meta.json` file being picked as a script's file path instead of the actual Luau file